[bug] check if account is new by checking accounts
previous behavior was using context (which screen we were on) to
determine if account was new or not.

in order to be able to have more straightforward automation this
commit now checks the list of accounts to get that information
which is more accurate and allows to login without navigating screens.